requirements.txt
fastapi
uvicorn
google-cloud-vision
google-generativeai